fix(core): extract CategoryPageHeader and FiltersSidebar
