+++ packages/core/docs/changelog/5684.js
@@ -1,0 +1,8 @@
+module.exports = {
+  description: 'Extract `CategoryPageHeader` and `FiltersSidebar` to separate components',
+  link: 'https://github.com/vuestorefront/vue-storefront/issues/5684',
+  isBreaking: false,
+  breakingChanges: [],
+  author: 'Łukasz Jędrasik',
+  linkToGitHubAccount: 'https://github.com/lukaszjedrasik'
+};
